Remove browser-template, add node-template
There's a strong consensus in #8 that this lesson uses
`browser-template` for historicaly reason and should be updated
to use `node-template`. This commit removes all files associated
with `browser-template` and adds everything from `node-template`.

The result of a `git diff --name-status` against `node-template`
at the time of this commit shows only `README.md` and `NOTES.md`.

Close #8. Close #47 because we won't be using JQuery anymore.
Close #58 since we won't be running code in the browser.
Close #71 because we won't be using Webpack and `this` binds
to window like we discuss on most websites.