Add discussion and lab for method invocation pattern
This is how `self` works in Ruby, so devs should be pretty familiar
with all this by now. So I think the previous code-along may have
been somewhat redundant. I changed it to a lab, where they see a
similar example of an object with a method that relies on `this`,
and they have to change the code so that method chaining works.