Add examples/exercises to `lib`
`lib/call.js` is an example of using `.call`. I moved it out so that
the instructor could show it and ask devs what they think will happen,
then run it as a demo.

`lib/bind.js` is a lab where devs need to use read the docs on `.bind`
and then use it in a simple example.